+++ packages/@expo/cli/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### 🎉 New features
 
+- Add support in `expo start` for custom app ids. ([#39236](https://github.com/expo/expo/pull/39236) by [@sebryu](https://github.com/sebryu))
+
 ### 🐛 Bug fixes
 
 - Prevent iOS device builds from being remotely cached ([#39621](https://github.com/expo/expo/pull/39621) by [@gabrieldonadel](https://github.com/gabrieldonadel))